+++ server/TWS Foundation/TWS Business.Quality/Depots/Q_TrucksCommonsDepot.cs
@@ -18,7 +18,6 @@ public class Q_TrucksCommonsDepot
 
         return new() {
             Status = 1,
-            Vin = RandomUtils.String(17),
             Economic = RandomUtils.String(16),
         };
     }